refactor: FestivalNotificationService 패키지 이동 및 API 요청 로직 개선
* **refactor: FestivalNotificationService 관련 구조 개선 및 오타 수정**
    * `FestivalNotificationService` 인터페이스를 `com.daedan.festabook.data.service.platform`에서 `com.daedan.festabook.data.service` 패키지로 이동했습니다.
    * `saveFestivalNotification` API 엔드포인트 경로의 중괄호 오타(`{platform}}` -> `{platform}`)를 수정했습니다.
    * `FestivalNotificationRemoteDataSourceTest`에서 `saveFestivalNotification` 호출 시 플랫폼 파라미터를 전달하도록 수정했습니다.

* **refactor: DeviceRepositoryImpl 기기 등록 로직 단순화**
    * `registerDevice` 메서드에서 FCM 토큰 존재 여부를 확인하던 조건부 로직을 제거하고, 전달받은 토큰을 항상 저장 및 사용하도록 로직을 단순화했습니다.
