+++ composeApp/src/commonMain/kotlin/com/daedan/festabook/data/service/FestivalNotificationService.kt
@@ -9,7 +9,7 @@ import de.jensklingenberg.ktorfit.http.POST
 import de.jensklingenberg.ktorfit.http.Path
 
 interface FestivalNotificationService {
-    @POST("festivals/{festivalId}/notifications/{platform}}")
+    @POST("festivals/{festivalId}/notifications/{platform}")
     suspend fun saveFestivalNotification(
         @Path("festivalId") id: Long,
         @Path("platform") platform: String,
